Fix warnings, remove debug messages

+++ base/plugins/scenario2/Document/Interval/IntervalContent/Storey/StoreyView.cpp
@@ -15,7 +15,7 @@ QRectF StoreyView::boundingRect() const
 	return {parentItem()->boundingRect().x() + DEMO_PIXEL_SPACING_TEST,
 			parentItem()->boundingRect().y() + 20,
 			parentItem()->boundingRect().width() - 2 * DEMO_PIXEL_SPACING_TEST,
-			m_height};
+			qreal(m_height)};
 }
 
 void StoreyView::paint(QPainter* painter, const QStyleOptionGraphicsItem* option, QWidget* widget)
